+++ lib/nfg_ui/components/elements/step.rb
@@ -9,20 +9,26 @@ module NfgUi
 
         include NfgUi::Components::Utilities::Iconable
 
+        include NfgUi::Components::Traits::Active
         include NfgUi::Components::Traits::Step
+
+        # Override #active from Activatable
+        def active
+          # Active is a transitory state for steps
+          # So when a step is active, it is automatically visited.
+          # This sets visited to true.
+          # This is also accounted for in the #active_trait for steps.
+          options[:visited] = true if options[:active]
+          super
+        end
 
         def component_family
           :steps
         end
 
-        def step
-          options.fetch(:step, nil)
-        end
-
-        def visited
-          options.fetch(:visited, false)
-        end
-
+        # NOTE: DISABLED TOOLTIPS DO NOT YET WORK ON STEPS
+        # NAVITEM DOES NOT INCLUDE DISABLEABLE MODULE
+        # JR: 06-21-2019
         def disabled
           options[:disabled] || (!visited && !active)
         end
@@ -31,6 +37,14 @@ module NfgUi
           content_tag(as, html_options) do
             NfgUi::Components::Elements::StepIndicator.new({ step: step.to_s, body: (block_given? ? yield : body), icon: icon, href: href, disabled: disabled }, view_context).render
           end
+        end
+
+        def step
+          options.fetch(:step, nil)
+        end
+
+        def visited
+          options.fetch(:visited, false)
         end
 
         private